refactor: merge world-market leftover test-support expectations (#4410)
Slice C folds treasury, validator, deal-matcher, FRR, suggester, and purchased-tile pin types into siblings under the 220-line cap. Slice D ratchets support LOC to 6590 and forbids new *_expectations.dart modules.
