+++ SPEC/program/repo-lint.md
@@ -243,8 +243,9 @@
 | `tool/check_economy_test_no_local_stockpile_game_builder.dart` | Forbid local `_buildGame` / `buildGame` / `_gameWith*` helpers under `packages/colonizethis_economy/test/**`; use shared builders from test_support (Refs #3831, #3856); rule `repo.economy_test_no_local_stockpile_game_builder` |
 | `tool/check_economy_bid_treasury_spend_shared.dart` | Require `trade_order_validator.dart` to import the `treasury_bid_budget.dart` barrel and call the shared `bidTreasurySpendForOrder` helper (defined in the phase-7 sibling `treasury_bid_spend.dart`) so per-bid treasury spend is not re-inlined (Refs #3823, #4049); rule `repo.economy_bid_treasury_spend_shared` |
 | `tool/check_economy_test_file_size.dart` | PR-blocking structural check: files under `packages/colonizethis_economy/test/**` must stay at or below 400 physical lines (mirrors `repo.logic_test_file_size`; Refs #3823); rule `repo.economy_test_file_size` |
-| `tool/check_economy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_economy_test_support/lib/**` physical Dart LOC (#4108 ≤7150, measured ≈6749 post-densify; phase-7 was ≤7495); rule `repo.economy_test_support_loc` (Refs #4014, #4049, #4108) |
+| `tool/check_economy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_economy_test_support/lib/**` physical Dart LOC (#4410 ≤6590, measured 6540 post leftover-merge; #4108 was ≤7150); rule `repo.economy_test_support_loc` (Refs #4014, #4049, #4108, #4410) |
 | `tool/check_economy_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_economy_test_support/lib/**` physical lines per file (#4108 ≤220, chosen just above the post-split largest file `deal_matcher_priority_scenarios.dart` ≈215); generated suffixes excluded; rule `repo.economy_test_support_file_size` (Refs #4108) |
+| `tool/check_economy_test_support_no_expectations_modules.dart` | Forbid any `*_expectations.dart` under `packages/colonizethis_economy_test_support/lib/**` after the leftover pin-type merge; rule `repo.economy_test_support_no_expectations_modules` (Refs #4410) |
 | `tool/check_combat_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_combat_test_support/lib/**` physical Dart LOC (#4196 ≤7250, measured ≈7242 post-split; baseline was 7,307); rule `repo.combat_test_support_loc` (Refs #4196) |
 | `tool/check_combat_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_combat_test_support/lib/**` physical lines per file (#4196 ≤220, chosen just above the post-split largest file `quick_battle_input_builder_scenarios.dart` = 220); generated suffixes excluded; rule `repo.combat_test_support_file_size` (Refs #4196) |
 | `tool/check_combat_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_combat/lib/**` physical lines per file (phase-4 ≤280, chosen just above the post-split largest file `quick_battle_resolver_engine.dart` ≈268; stricter than the shared 550-line `repo.domain_package_source_file_size` cap); generated suffixes excluded; shrink-only grandfather allowlist during transition; rule `repo.combat_lib_file_size` (Refs #4284) |
@@ -581,6 +582,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a non-support `packages/colonizethis_world/test/**` Dart file with more than 320 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.colonizethis_world_test_file_size`, then the run fails naming that file and exits `1` (Refs #4125, #4330).
 - Given every non-generated Dart file under `packages/colonizethis_turn/lib/**` has ≤300 physical lines and the shrink-only grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.turn_lib_file_size`, then the rule passes without violations and exits `0` (Refs #4113, #4342).
 - Given a non-generated `packages/colonizethis_turn/lib/**` Dart file with more than 300 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.turn_lib_file_size`, then the run fails naming that file and exits `1` (Refs #4113, #4342).
+- Given `packages/colonizethis_economy_test_support/lib/**` has no `*_expectations.dart` files, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_test_support_no_expectations_modules`, then the rule passes without violations and exits `0` (Refs #4410).
+- Given a planted `packages/colonizethis_economy_test_support/lib/src/foo_expectations.dart`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_test_support_no_expectations_modules`, then the run fails listing that file and exits `1` (Refs #4410).
 - Given the `colonizethis_economy` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_no_part_directives`, then the rule passes without violations and exits `0` (Refs #3979).
 - Given a `packages/colonizethis_economy/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.economy_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #3979).
 - Given the `colonizethis_models` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4068).
