+++ packages/colonizethis_economy_test_support/lib/colonizethis_economy_test_support.dart
@@ -40,7 +40,6 @@ export 'src/projected_cost_engine_scenarios.dart';
 export 'src/province_extraction_snapshot_pins.dart';
 export 'src/province_extraction_snapshot_scenarios.dart';
 export 'src/purchased_tile_fixture_support.dart';
-export 'src/purchased_tile_scenarios/purchased_tile_expectations.dart';
 export 'src/purchased_tile_scenarios/purchased_tile_scenarios.dart';
 export 'src/town_manufacturing_bonus_scenarios.dart';
 export 'src/trade_interception_scenarios.dart';
@@ -48,7 +47,6 @@ export 'src/worker_action_cost_scenarios.dart';
 export 'src/worker_economy_scenarios.dart';
 export 'src/trade_order_admission_scenarios.dart';
 export 'src/trade_order_factory.dart';
-export 'src/trade_order_suggester_expectations.dart';
 export 'src/trade_order_suggester_scenarios.dart';
 export 'src/trade_order_suggester_test_support.dart';
 export 'src/validator_scenarios/validator_scenarios.dart';
