+++ packages/colonizethis_economy_test_support/lib/src/deal_matcher_scenarios/deal_matcher_frr_scenarios.dart
@@ -1,7 +1,6 @@
 // FRR DealMatcher scenarios (Refs #3836, #3939).
 import 'package:colonizethis_economy/colonizethis_economy.dart';
 import '../frr_scenarios/frr_d5_test_support.dart';
-import 'deal_matcher_expectations.dart';
 import 'deal_matcher_scenario.dart';
 import 'deal_matcher_test_support.dart';
