+++ packages/colonizethis_economy_test_support/lib/src/deal_matcher_scenarios/deal_matcher_priority_scenarios.dart
@@ -1,6 +1,5 @@
 // dart format off
 // Table-driven DealMatcher scenarios (Refs #3836, #3939).
-import 'deal_matcher_expectations.dart';
 import 'deal_matcher_scenario.dart';
 import 'deal_matcher_test_support.dart';
 /// Priority and FTP precedence from `world_market_deal_matcher_priority_test.dart`.
