+++ packages/colonizethis_economy_test_support/lib/src/deal_matcher_scenarios/deal_matcher_scenarios.dart
@@ -1,7 +1,6 @@
 // Barrel for consolidated DealMatcher scenario modules (Refs #3939 phase 3 slice 7).
 
 export 'deal_matcher_basic_scenarios.dart';
-export 'deal_matcher_expectations.dart';
 export 'deal_matcher_frr_scenarios.dart';
 export 'deal_matcher_priority_scenarios.dart';
 export 'deal_matcher_scenario.dart';
