+++ packages/colonizethis_economy_test_support/lib/src/deal_matcher_scenarios/deal_matcher_treasury_scenarios.dart
@@ -1,6 +1,5 @@
 // Table-driven DealMatcher scenarios (Refs #3836, #3939).
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'deal_matcher_expectations.dart';
 import 'deal_matcher_scenario.dart';
 import 'deal_matcher_test_support.dart';
