+++ packages/colonizethis_economy_test_support/lib/src/frr_scenarios/frr_scenarios.dart
@@ -1,7 +1,6 @@
 // Barrel for consolidated First Right of Refusal scenario modules (Refs #3939 phase 3 slice 7).
 
 export 'first_right_profit_scenarios.dart';
-export 'frr_credits_expectations.dart';
 export 'frr_credits_scenarios.dart';
 export 'frr_credits_test_support.dart';
 export 'frr_d5_test_support.dart';
