+++ packages/colonizethis_economy_test_support/lib/src/purchased_tile_scenarios/purchased_tile_index_scenarios.dart
@@ -4,7 +4,6 @@ import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/game_test_fixtures.dart';
 import 'package:colonizethis_test/test.dart';
-import 'purchased_tile_expectations.dart';
 import 'purchased_tile_index_test_support.dart';
 /// One row in [purchasedTileAttributionSemanticsScenarios] (Refs #3979).
 typedef PurchasedTileAttributionSemanticsScenario = ({
