+++ packages/colonizethis_economy_test_support/lib/src/purchased_tile_scenarios/purchased_tile_index_test_support.dart
@@ -1,6 +1,8 @@
 // dart format off
 /// Shared fixtures for [PurchasedTileIndex] unit tests (Refs #3856, #3939).
+import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:colonizethis_test/test.dart';
 import '../purchased_tile_fixture_support.dart';
 export '../purchased_tile_fixture_support.dart' show purchasedTileFixtureGame, minorPurchasedTileGame;
 /// Canonical minor-owned purchased-tile scenario used by AC-D1-2 and AC-D1-7.
@@ -65,5 +67,92 @@ Game mixedMinorTribePurchasedTileGame() {
 Game emptyOwningGpPurchasedTileGame() {
   const tileKey = 'oldWorld|M1|0|0';
   return minorPurchasedTileGame(minorDisplayName: 'Minor 1', purchasedTilesByTileKey: const {tileKey: ''});
+}
+/// Data-driven expectations for [PurchasedTileIndexFromGameScenario] rows.
+class PurchasedTileIndexExpectation {
+  const PurchasedTileIndexExpectation({this.length, this.isEmpty, this.isNotEmpty, this.attributionForTileKeyNull, this.attributionForTileKeysNull, this.attributionsEmpty = false, this.singleAttribution, this.multiAttributions, this.attributionsTileKeysContainAll, this.deterministicIndexRerun = false, this.deterministicRerunTileKey});
+  final int? length;
+  final bool? isEmpty;
+  final bool? isNotEmpty;
+  final String? attributionForTileKeyNull;
+  final Iterable<String>? attributionForTileKeysNull;
+  final bool attributionsEmpty;
+  final PurchasedTileAttributionExpectation? singleAttribution;
+  final Iterable<PurchasedTileAttributionExpectation>? multiAttributions;
+  final Iterable<String>? attributionsTileKeysContainAll;
+  final bool deterministicIndexRerun;
+  final String? deterministicRerunTileKey;
+}
+class PurchasedTileAttributionExpectation {
+  const PurchasedTileAttributionExpectation({required this.tileKey, this.owningGpId, this.sourceFactionId, this.provinceId});
+  final String tileKey;
+  final String? owningGpId;
+  final String? sourceFactionId;
+  final String? provinceId;
+}
+void assertPurchasedTileIndexExpectation(PurchasedTileIndex index, PurchasedTileIndexExpectation expectation, {Game? gameForDeterminismRerun}) {
+  if (expectation.length != null) {
+    expect(index.length, expectation.length);
+  }
+  if (expectation.isEmpty != null) {
+    expect(index.isEmpty, expectation.isEmpty);
+  }
+  if (expectation.isNotEmpty != null) {
+    expect(index.isNotEmpty, expectation.isNotEmpty);
+  }
+  if (expectation.attributionForTileKeyNull != null) {
+    expect(index.attributionForTileKey(expectation.attributionForTileKeyNull!), isNull);
+  }
+  if (expectation.attributionForTileKeysNull != null) {
+    for (final tileKey in expectation.attributionForTileKeysNull!) {
+      expect(index.attributionForTileKey(tileKey), isNull);
+    }
+  }
+  if (expectation.attributionsEmpty) {
+    expect(index.attributions, isEmpty);
+  }
+  if (expectation.singleAttribution != null) {
+    final attr = index.attributionForTileKey(expectation.singleAttribution!.tileKey);
+    expect(attr, isNotNull);
+    if (expectation.singleAttribution!.owningGpId != null) {
+      expect(attr!.owningGpId, expectation.singleAttribution!.owningGpId);
+    }
+    if (expectation.singleAttribution!.sourceFactionId != null) {
+      expect(attr!.sourceFactionId, expectation.singleAttribution!.sourceFactionId);
+    }
+    if (expectation.singleAttribution!.provinceId != null) {
+      expect(attr!.provinceId, expectation.singleAttribution!.provinceId);
+    }
+    expect(attr!.tileKey, expectation.singleAttribution!.tileKey);
+  }
+  if (expectation.multiAttributions != null) {
+    for (final pin in expectation.multiAttributions!) {
+      final attr = index.attributionForTileKey(pin.tileKey);
+      expect(attr, isNotNull);
+      if (pin.owningGpId != null) {
+        expect(attr!.owningGpId, pin.owningGpId);
+      }
+      if (pin.sourceFactionId != null) {
+        expect(attr!.sourceFactionId, pin.sourceFactionId);
+      }
+      if (pin.provinceId != null) {
+        expect(attr!.provinceId, pin.provinceId);
+      }
+    }
+  }
+  if (expectation.attributionsTileKeysContainAll != null) {
+    expect(index.attributions.map((a) => a.tileKey).toSet(), containsAll(expectation.attributionsTileKeysContainAll!));
+  }
+  if (expectation.deterministicIndexRerun) {
+    final game = gameForDeterminismRerun;
+    expect(game, isNotNull, reason: 'deterministicIndexRerun requires game');
+    final first = PurchasedTileIndex.fromGame(game!);
+    final second = PurchasedTileIndex.fromGame(game);
+    expect(index.length, first.length);
+    expect(first.length, second.length);
+    if (expectation.deterministicRerunTileKey != null) {
+      expect(first.attributionForTileKey(expectation.deterministicRerunTileKey!), equals(second.attributionForTileKey(expectation.deterministicRerunTileKey!)));
+    }
+  }
 }
 // dart format on
