+++ packages/colonizethis_economy_test_support/lib/src/treasury_scenarios/treasury_available_scenarios.dart
@@ -4,7 +4,6 @@ import 'package:colonizethis_data/colonizethis_data.dart' as data;
 import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
-import 'treasury_expectations.dart';
 import 'treasury_test_support.dart';
 /// One row for [treasuryAvailableForBidsScenarios].
 typedef TreasuryAvailableScenario = ({String label, int treasury, String playerId, int projectedNonBidTreasuryDelta, int expected, String? refs, TreasuryAvailableExpectation? extra});
