+++ packages/colonizethis_economy_test_support/lib/src/validator_scenarios/validator_rules_scenarios.dart
@@ -1,7 +1,6 @@
 // Rules 1–3 and empty/accept validator scenarios (Refs #3856, #3939 phase 3 slice 30).
 import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'trade_order_validator_test_support.dart';
-import 'validator_expectations.dart';
 import 'validator_scenario.dart';
 /// Rules 1–3 and empty/accept paths from
 /// `world_market_trade_order_validator_test.dart` (Refs #3856 slice 8).
